+++ lib/cancelable_http_client.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:http/http.dart';
+import 'package:multitasking/misc/pause.dart';
 import 'package:multitasking/multitasking.dart';
 
 export 'package:multitasking/multitasking.dart'
@@ -30,15 +31,33 @@ export 'package:multitasking/multitasking.dart'
 class CancelableClient with BaseClient {
   final Client _client;
 
+  final PauseToken? _pauseToken;
+
+  final Duration? _requestTimeout;
+
+  final Duration? _responseTimeout;
+
   final CancellationToken _token;
 
   /// Creates an instance of [CancelableClient].\
   ///
   /// Parameters:
   ///
+  /// - [pauseToken]: Token for pausing and resuming data retrieval from the
+  /// server.
+  /// - [requestTimeout]: Time limit for waiting for a response from the server.
+  /// - [responseTimeout: Time limit for waiting for data to be received from
+  /// the server (maximum allowed delay between `onData` events).
   /// - [token]: Cancellation token to signal a cancellation request.
-  CancelableClient(CancellationToken token)
-      : _client = Client(),
+  CancelableClient(
+    CancellationToken token, {
+    PauseToken? pauseToken,
+    Duration? requestTimeout,
+    Duration? responseTimeout,
+  })  : _client = Client(),
+        _pauseToken = pauseToken,
+        _requestTimeout = requestTimeout,
+        _responseTimeout = responseTimeout,
         _token = token;
 
   @override
@@ -50,28 +69,53 @@ class CancelableClient with BaseClient {
   Future<StreamedResponse> send(BaseRequest request) async {
     _token.throwIfCanceled();
     final task = Task.run(() => _client.send(request));
+    final tasks = [task.withCancellation(_token)];
+    Timer? timer;
+    if (_requestTimeout != null) {
+      final completer = TaskCompletionSource<StreamedResponse>();
+      timer = Timer(_requestTimeout!, () {
+        completer.setError(TimeoutException(''), StackTrace.current);
+      });
+
+      tasks.add(completer.task);
+    }
+
+    Future<void> cancelRequest() async {
+      try {
+        // Wait for a response from the server.
+        final response = await task;
+        final stream = response.stream;
+        // Notify the server to cancel the data transfer.
+        await stream.listen(null).cancel();
+      } catch (e) {
+        // Ignore exception
+      }
+    }
+
     StreamedResponse response;
     try {
-      response = await task.withCancellation(_token);
+      response = await Future.any(tasks);
     } on CancellationException {
-      // Ignore the cancelled operation.
-      unawaited(() async {
-        try {
-          // Wait for a response from the server.
-          final response = await task;
-          final stream = response.stream;
-          // Notify the server to cancel the data transfer.
-          await stream.listen(null).cancel();
-        } catch (e) {
-          // Ignore exception
-        }
-      }());
-
+      timer?.cancel();
+      unawaited(cancelRequest());
       rethrow;
+    } on TimeoutException {
+      unawaited(cancelRequest());
+      rethrow;
+    } finally {
+      timer?.cancel();
     }
 
     final stream = response.stream;
-    final cancelableStream = stream.asCancelable(_token);
+    var cancelableStream = stream.asCancelable(
+      _token,
+      timeout: _responseTimeout,
+    );
+
+    if (_pauseToken != null) {
+      cancelableStream = cancelableStream.asPausable(_pauseToken!);
+    }
+
     return StreamedResponse(
       cancelableStream,
       response.statusCode,